Refactored codebase
