+++ src/main/scala/linting/LintException.scala
@@ -5,20 +5,25 @@ package freechips.rocketchip.linting
 import firrtl.FirrtlUserException
 import firrtl.ir.FileInfo
 
-object LintExceptions {
-  def makeNumber(max: Int, n: Int, prefix: String): String = {
+case class LintException(seq: Seq[Violation], lintDisplayOptions: DisplayOptions) extends FirrtlUserException(
+  LintException.buildMessage(seq, lintDisplayOptions)
+)
+
+object LintException {
+  private def makeNumber(max: Int, n: Int, prefix: String): String = {
     val nDigits = max.toString.size
     val nLeading = nDigits - n.toString.size
     prefix * nLeading + n.toString
   }
-  private[linting] def buildMessage(seq: Seq[LintError], lintDisplayOptions: LintDisplayOptions): String = {
+
+  private[linting] def buildMessage(seq: Seq[Violation], lintDisplayOptions: DisplayOptions): String = {
     val groupedErrors = seq.groupBy {
-      case l: LintError => l.linter.lintNumber
+      case l: Violation => l.linter.lintNumber
     }
     val maxErrorNumber = groupedErrors.keys.max
 
     val (_, reports) = groupedErrors.toSeq.sortBy(_._1).reverse.foldRight((0, Seq.empty[String])) {
-      case ((lintNumber: Int, lintErrors: Seq[LintError]), (totalErrors: Int, reportsSoFar: Seq[String])) =>
+      case ((lintNumber: Int, lintErrors: Seq[Violation]), (totalErrors: Int, reportsSoFar: Seq[String])) =>
         val le = lintErrors.head.linter
         val perErrorLimit = lintDisplayOptions.perErrorLimit.getOrElse(lintNumber, lintErrors.size)
         val totalErrorLimit = lintDisplayOptions.totalLimit.map(t => t - totalErrors).getOrElse(perErrorLimit)
@@ -37,7 +42,7 @@ object LintExceptions {
              | - Disable this linting check:
              |     ${le.disableCLI}
              |${lintErrors.zip(1 to actualNumErrors)
-                   .map { case (lint: LintError, idx: Int) => s"#$lintString.${makeNumber(actualNumErrors - 1,idx,"0")}:${lint.info} ${lint.message} in ${lint.modules}"}
+                   .map { case (lint: Violation, idx: Int) => s"#$lintString.${makeNumber(actualNumErrors - 1,idx,"0")}:${lint.info} ${lint.message} in ${lint.modules}"}
                    .mkString("\n")}""".stripMargin
         (totalErrors + actualNumErrors, report +: reportsSoFar)
     }
@@ -45,7 +50,4 @@ object LintExceptions {
   }
 }
 
-case class LintExceptions(seq: Seq[LintError], lintDisplayOptions: LintDisplayOptions) extends FirrtlUserException(
-  LintExceptions.buildMessage(seq, lintDisplayOptions)
-)
