+++ src/main/scala/linting/rule/LintAnonymousRegisters.scala
@@ -1,20 +1,16 @@
 // See LICENSE.SiFive for license details.
 
 package freechips.rocketchip.linting
+package rule
 
-import firrtl._
 import firrtl.ir._
-import firrtl.traversals.Foreachers._
-import firrtl.annotations.NoTargetAnnotation
-import firrtl.options.{OptionsException, RegisteredLibrary, ShellOption, Dependency}
-import firrtl.stage.RunFirrtlTransformAnnotation
-import chisel3.experimental.ChiselAnnotation
-import scala.collection.mutable
+import firrtl.options.Dependency
 
 
-final class LintAnonymousRegisters extends Linter {
-  // Import useful utility functions
-  import Linter.{isTemporary, getName, getScalaInfo, isWhitelisted, Errors, updateErrors}
+/** Reports all anonymous registers in design
+  * An anonymous register is one which is prefixed with an "_"
+  */
+final class LintAnonymousRegisters extends LintRule {
 
   val recommendedFix: String = "Use named intermediate val, or if that fails use @chiselName or *.suggestName(...)"
 
@@ -22,14 +18,16 @@ final class LintAnonymousRegisters extends Linter {
 
   val lintName: String = "anon-regs"
 
+  // Should run before LowerTypes so anonymous aggregate registers are reported as one register
   override def optionalPrerequisiteOf = super.optionalPrerequisiteOf :+ Dependency(firrtl.passes.LowerTypes)
 
   override protected def lintStatement(errors: Errors, mname: String)(s: Statement): Unit = {
     s match {
       case r: DefRegister  if isTemporary(r.name) =>
+        // Report scala info, if its present. Otherwise, use existing Info
         getScalaInfo(r.info) match {
           case Some(scalaInfo: FileInfo) => updateErrors(scalaInfo, "", errors, mname)
-          case None => updateErrors(r.info, "", errors, mname)
+          case None                      => updateErrors(r.info, "", errors, mname)
         }
       case other =>
     }
